+++ tests/Firebase/Database/Query/Sorter/OrderByChildTest.php
@@ -11,44 +11,30 @@ class OrderByChildTest extends FirebaseTestCase
     /**
      * @var OrderByChild
      */
-    protected $ascending;
-
-    /**
-     * @var OrderByChild
-     */
-    protected $descending;
+    protected $sorter;
 
     protected function setUp()
     {
-        $this->ascending = new OrderByChild('key');
-        $this->descending = new OrderByChild('key', SORT_DESC);
+        $this->sorter = new OrderByChild('key');
     }
 
     public function testModifyUri()
     {
         $this->assertContains(
             'orderBy='.rawurlencode('"key"'),
-            (string) $this->ascending->modifyUri(new Uri('http://domain.tld'))
+            (string) $this->sorter->modifyUri(new Uri('http://domain.tld'))
         );
     }
 
     /**
-     * @dataProvider sortAscendingValueProvider
+     * @dataProvider valueProvider
      */
-    public function testSortAscending($expected, $value)
+    public function testModifyValue($expected, $value)
     {
-        $this->assertSame($expected, $this->ascending->modifyValue($value));
+        $this->assertSame($expected, $this->sorter->modifyValue($value));
     }
 
-    /**
-     * @dataProvider sortDescendingValueProvider
-     */
-    public function testSortDescending($expected, $value)
-    {
-        $this->assertSame($expected, $this->descending->modifyValue($value));
-    }
-
-    public function sortAscendingValueProvider()
+    public function valueProvider()
     {
         return [
             'scalar' => [
@@ -61,30 +47,6 @@ class OrderByChildTest extends FirebaseTestCase
                     'fourth' => ['key' => 2],
                     'first' => ['key' => 3],
                     'second' => ['key' => 4],
-                ],
-                'given' => [
-                    'first' => ['key' => 3],
-                    'second' => ['key' => 4],
-                    'third' => ['key' => 1],
-                    'fourth' => ['key' => 2],
-                ],
-            ],
-        ];
-    }
-
-    public function sortDescendingValueProvider()
-    {
-        return [
-            'scalar' => [
-                'expected' => 'scalar',
-                'given' => 'scalar',
-            ],
-            'array' => [
-                'expected' => [
-                    'second' => ['key' => 4],
-                    'first' => ['key' => 3],
-                    'fourth' => ['key' => 2],
-                    'third' => ['key' => 1],
                 ],
                 'given' => [
                     'first' => ['key' => 3],